Refactor repository interface module
